requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
SQLAlchemy==2.0.32
psycopg[binary]==3.2.10
alembic==1.13.2
celery==5.4.0
redis==5.0.7
minio==7.2.8
boto3==1.34.162
python-multipart==0.0.9
markdown-it-py==3.0.0
Pillow==10.4.0
opencv-python-headless==4.10.0.84
pytesseract==0.3.13
langdetect==1.0.9
orjson==3.10.7
tenacity==8.5.0
structlog==24.1.0
requests==2.32.3
pytest==8.3.2
httpx==0.27.2
prometheus-client==0.20.0
pypdf==4.2.0
